refactor: move login&registration helpers to lib folder
