+++ src/lib/auth/validation.ts
@@ -1,5 +1,14 @@
 import { z } from 'zod';
 
+export const loginSchema = z.object({
+	email: z.string().email({
+		message: 'Email must be a valid email address',
+	}),
+	password: z.string().min(6, {
+		message: 'Password must be at least 6 characters long',
+	}),
+});
+
 export const registrationSchema = z.object({
 	firstName: z.string().min(2, {
 		message: 'First name must be at least 2 characters long',
